Fix 64-bit compilation issues and implement Moveable<> inheritance rules
This commit addresses several important compatibility issues:

1. Fixed 64-bit compilation problems by:
   - Changing pointer-to-integer casts to use proper types like ptrdiff_t
   - Updating printf format specifiers from %x to %p for pointer values
   - Adding necessary std::string includes for C++ standard library usage

2. Implemented Moveable<> inheritance rules by:
   - Making classes that need move semantics inherit from Moveable<>
   - Ensuring Vector template parameters properly inherit from Moveable<>
   - Following proper assignment rules for Vector instances

These changes improve code stability, maintainability, and compatibility with 64-bit systems.


Fixes include updates to ratl_common.h, snd_mem.cpp, snd_dma.cpp, snd_ambient.cpp, snd_music.cpp, ai_main.c, bg_public.h, sof2mp_types.h, and adding QWEN.md for project guidelines.

+++ code/Ratl/ratl_common.h
@@ -590,7 +590,7 @@ namespace storage
 		}
 		static int pointer_to_index(const void *s1,const void *s2)
 		{
-			return ((TStorage *)s1)-((TStorage *)s2);
+			return (int)(((TStorage *)s1)-((TStorage *)s2));
 		}
 	};
 	template<class T,int SIZE>
@@ -658,7 +658,7 @@ namespace storage
 		}
 		static int pointer_to_index(const void *s1,const void *s2)
 		{
-			return ((TStorage *)s1)-((TStorage *)s2);
+			return (int)(((TStorage *)s1)-((TStorage *)s2));
 		}	
 	};
 	template<class T,int SIZE,int MAX_CLASS_SIZE>
@@ -722,7 +722,7 @@ namespace storage
 		}
 		static int pointer_to_index(const void *s1,const void *s2)
 		{
-			return ((TStorage *)s1)-((TStorage *)s2);
+			return (int)(((TStorage *)s1)-((TStorage *)s2));
 		}	
 		template<class CAST_TO>
 		static CAST_TO *verify_alloc(CAST_TO *p)
@@ -805,11 +805,11 @@ namespace storage
 		// this is so node support does not need to be added to the primitive containers
 		static NODE & node(TValue &v)
 		{
-			return *(NODE *)((unsigned char *)(&v)+int(&((TStorage *)0)->nodeData)-int(&((TStorage *)0)->value));
+			return *(NODE *)((unsigned char *)(&v)+(ptrdiff_t)(&((TStorage *)0)->nodeData)-(ptrdiff_t)(&((TStorage *)0)->value));
 		}
 		static const NODE & node(const TValue &v)
 		{
-			return *(const NODE *)((unsigned char *)(&v)+int(&((TStorage *)0)->nodeData)-int(&((TStorage *)0)->value));
+			return *(const NODE *)((unsigned char *)(&v)+(ptrdiff_t)(&((TStorage *)0)->nodeData)-(ptrdiff_t)(&((TStorage *)0)->value));
 		}
 		static void swap(TStorage *s1,TStorage *s2)
 		{
@@ -819,8 +819,8 @@ namespace storage
 		static int pointer_to_index(const void *s1,const void *s2)
 		{
 			return 
-				((TStorage *)(((unsigned char *)s1)-int(&((TStorage *)0)->value))) - 
-				((TStorage *)(((unsigned char *)s2)-int(&((TStorage *)0)->value)));
+				((TStorage *)(((unsigned char *)s1)-(ptrdiff_t)(&((TStorage *)0)->value))) - 
+				((TStorage *)(((unsigned char *)s2)-(ptrdiff_t)(&((TStorage *)0)->value)));
 		}	
 	};
 
@@ -900,11 +900,11 @@ namespace storage
 		// this is so node support does not need to be added to the primitive containers
 		static NODE & node(TValue &v)
 		{
-			return *(NODE *)((unsigned char *)(&v)+int(&((TStorage *)0)->nodeData)-int(&((TStorage *)0)->value));
+			return *(NODE *)((unsigned char *)(&v)+(ptrdiff_t)(&((TStorage *)0)->nodeData)-(ptrdiff_t)(&((TStorage *)0)->value));
 		}
 		static const NODE & node(const TValue &v)
 		{
-			return *(const NODE *)((unsigned char *)(&v)+int(&((TStorage *)0)->nodeData)-int(&((TStorage *)0)->value));
+			return *(const NODE *)((unsigned char *)(&v)+(ptrdiff_t)(&((TStorage *)0)->nodeData)-(ptrdiff_t)(&((TStorage *)0)->value));
 		}
 		static void swap(TStorage *s1,TStorage *s2)
 		{
@@ -916,8 +916,8 @@ namespace storage
 		static int pointer_to_index(const void *s1,const void *s2)
 		{
 			return 
-				((TStorage *)(((unsigned char *)s1)-int(&((TStorage *)0)->value))) - 
-				((TStorage *)(((unsigned char *)s2)-int(&((TStorage *)0)->value)));
+				((TStorage *)(((unsigned char *)s1)-(ptrdiff_t)(&((TStorage *)0)->value))) - 
+				((TStorage *)(((unsigned char *)s2)-(ptrdiff_t)(&((TStorage *)0)->value)));
 		}	
 	};
 	template<class T,int SIZE,int MAX_CLASS_SIZE,class NODE>
@@ -991,11 +991,11 @@ namespace storage
 		// this is so node support does not need to be added to the primitive containers
 		static NODE & node(TValue &v)
 		{
-			return *(NODE *)((unsigned char *)(&v)+int(&((TStorage *)0)->nodeData)-int(&((TStorage *)0)->value));
+			return *(NODE *)((unsigned char *)(&v)+(ptrdiff_t)(&((TStorage *)0)->nodeData)-(ptrdiff_t)(&((TStorage *)0)->value));
 		}
 		static const NODE & node(const TValue &v)
 		{
-			return *(const NODE *)((unsigned char *)(&v)+int(&((TStorage *)0)->nodeData)-int(&((TStorage *)0)->value));
+			return *(const NODE *)((unsigned char *)(&v)+(ptrdiff_t)(&((TStorage *)0)->nodeData)-(ptrdiff_t)(&((TStorage *)0)->value));
 		}
 		// this is a bit suspicious, we are forced to do a memory swap, and for a class, that, say
 		// stores a pointer to itself, it won't work right
@@ -1007,8 +1007,8 @@ namespace storage
 		static int pointer_to_index(const void *s1,const void *s2)
 		{
 			return 
-				((TStorage *)(((unsigned char *)s1)-int(&((TStorage *)0)->value))) - 
-				((TStorage *)(((unsigned char *)s2)-int(&((TStorage *)0)->value)));
+				((TStorage *)(((unsigned char *)s1)-(ptrdiff_t)(&((TStorage *)0)->value))) - 
+				((TStorage *)(((unsigned char *)s2)-(ptrdiff_t)(&((TStorage *)0)->value)));
 		}	
 		template<class CAST_TO>
 		static CAST_TO *verify_alloc(CAST_TO *p)
